requirements.txt
python_telegram_bot==13.6
psycopg2==2.8.6
python-dotenv==0.17.1
pytz>=2013b
requests~=2.25.1